Create version branch
TL;DR
-----

Add a branch for tracking semantic versioning from Concourse

Details
-------

The Cloud Pipelines use the `semver` resource for Concourse to
track the version of the code in branch in the Git repo. This
commit creates that branch and makes it available to the pipeline.

To keep things tidy, I deleted all other files and updated the
`.gitignore` so that only the file `version` is allow on this
branch.